Refactor: Avoid needlessly querying BTC height at each block
