+++ contracts/btc-staking/src/state/mod.rs
@@ -4,8 +4,3 @@ pub mod public_randomness;
 pub mod staking;
 
 mod fp_index;
-
-use cw_storage_plus::Map;
-
-/// Map of BTC height by block height
-pub(crate) const BTC_HEIGHT: Map<u64, u64> = Map::new("btc_height");
